+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add taskfile.dev development environment into the API repository itself, closes #461.
 - Add S3 status check to healthcheck command, closes #200.
 - Add support to delete files in S3 / object storage through the `php bin/console database:drop` command, closes #309.
+- Add support to back up files in S3 / object storage through the `php bin/console backup:create` command, closes #308.
 - Add support for `Content-Disposition` HTTP header, including transliteration of non-latin-characters, closes #447.
 - Add dependency to PHP library `cardinalby/content-disposition`.
 - Add 'reserved type' exception; nodes of type 'User', 'Token' and 'Upload' can not be manually created through generic